src-tauri/pyproject.toml
[project]
name = "gera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "you" }
]
requires-python = ">=3.9"
dependencies = [
    "pytauri == 0.8.*",
    "pydantic == 2.*",
    "anyio == 4.*",
    "mistune >= 3.0",
    "pyyaml >= 6.0",
    "watchfiles >= 1.0",
]

[project.entry-points.pytauri]
ext_mod = "gera.ext_mod"

[build-system]
requires = ["setuptools >= 80", "setuptools-scm >= 8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = { where = ["src-python"] }
